Transition all containers to TypeScript
