+++ package.json
@@ -123,6 +123,10 @@
     "@types/react-redux": "^7.0.1",
     "@types/react-router-dom": "^4.3.1",
     "@types/react-tooltip": "^3.9.0",
+    "@types/redux-devtools": "^3.0.46",
+    "@types/redux-devtools-dock-monitor": "^1.1.33",
+    "@types/redux-devtools-log-monitor": "^1.0.34",
+    "@types/screenfull": "^4.1.0",
     "@types/seed-random": "^2.2.0",
     "@types/shuffle-array": "^0.0.28",
     "@types/webpack": "^4.0.0",
