+++ types/packages.d.ts
@@ -3,5 +3,6 @@ declare module 'prop-types';
 declare module 'react-bar-chart';
 declare module 'react-textfit';
 declare module 'react-tooltip';
+declare module 'screenfull';  // Not sure why TS isn't picking up the @types/screenfull package
 declare module 'seed-shuffle';
 declare module 'trianglify';
